Moved fragments into "fragments" package
