+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FF00FF"
-    tools:context=".activities.ProfileFragment">
+    tools:context=".fragments.ProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
